[python-frontend] reorganize preprocessor into a dedicated package
